Move functions in `Delta` into new `DeltaCollection`
`Delta` should represent a `Delta` object, and instead these functions should move into a 'Collection' class instead
